aider: Updated client-side JavaScript to always display elapsed time to user in both success and error cases.
# Aider chat conversation:

USER: explain """                queue.append((next, path + [next], depth + 1))
""" why isnt it enough to just store `next` in the queue? (dont change the code, just explain)

ASSISTANT: The line `queue.append((next, path + [next], depth + 1))` is adding a tuple to the queue. This tuple contains three elements:

1. `next`: This is the next page to be visited.

2. `path + [next]`: This is the path taken to reach the `next` page. It's a list of all the pages visited so far, in the order they were visited. This is necessary because the goal of the algorithm is not just to reach the finish page, but to find the path to the finish page. By keeping track of the path, the algorithm can return the sequence of pages that leads from the start page to the finish page.

3. `depth + 1`: This is the number of steps taken to reach the `next` page from the start page. This is also known as the depth of the `next` page in the breadth-first search tree. This information could be useful for analysis or debugging purposes, or if you want to limit the search to a certain depth.

So, while it would be possible to only store `next` in the queue, doing so would lose the information about the path and the depth, which are needed to achieve the goals of the algorithm.

USER: ok, i am happy now. can we come back to displaying the elapsed time to the user.

ASSISTANT: Sure, let's ensure that the elapsed time is displayed to the user in all cases.

From your previous changes, it seems like you have already added the elapsed time to the server response in both success and error cases. Now, we need to make sure that this information is displayed on the client side.

To do this, we need to modify `client/main.js`:

1. Update the success case to always display the elapsed time.
2. Update the error case to display the elapsed time if it's available.

Here are the *SEARCH/REPLACE* blocks:

client/main.js
```javascript
<<<<<<< SEARCH
        if (!data.path.length) {
            logsHtml += '<p>' + data.time + '</p>';
        }
=======
        logsHtml += '<p>Elapsed time: ' + data.time + '</p>';
>>>>>>> REPLACE
```

client/main.js
```javascript
<<<<<<< SEARCH
        logsElement.innerHTML = error.message;
=======
        logsElement.innerHTML = error.message + (data && data.time ? '<p>Elapsed time: ' + data.time + '</p>' : '');
>>>>>>> REPLACE
```

These changes will ensure that the elapsed time is always displayed to the user, regardless of whether the path finding was successful or not.